Allow search document ID
