+++ search-esdb-service/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"search-esdb-service/config"
 	"search-esdb-service/database"
 	recordMigrator "search-esdb-service/record/migration"
@@ -10,7 +9,6 @@ import (
 
 func main() {
 	cfg := config.GetConfig()
-	fmt.Println("CONFIG--------",cfg)
 	db := database.NewElasticDatabase(&cfg)
 	recordMigrator.RecordMigrate(&cfg, db)
 	server.NewGinServer(&cfg, db.GetDB()).Start()
