+++ search-esdb-service/record/repositories/recordESRepositoryInsert.go
@@ -1,0 +1,115 @@
+package repositories
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"log"
+	"search-esdb-service/record/entities"
+	"strconv"
+	"strings"
+	"sync/atomic"
+	"time"
+
+	"github.com/dustin/go-humanize"
+	"github.com/elastic/go-elasticsearch/v8/esutil"
+)
+
+// BulkInsert inserts multiple records into the Elasticsearch index.
+//
+// qars: A slice of pointers to Record entities representing the records to be inserted.
+// Returns an error if there was an issue inserting the records.
+func (r *RecordESRepository) BulkInsert(qars []*entities.Record) error {
+	es := r.es
+	var countSuccessful uint64
+	bi, err := esutil.NewBulkIndexer(esutil.BulkIndexerConfig{
+		Index:  "record",
+		Client: es, // The Elasticsearch client
+	})
+	if err != nil {
+		return fmt.Errorf("Error creating the indexer: %s", err)
+	}
+
+	start := time.Now().UTC()
+
+	// Loop over the collection
+	for order, a := range qars {
+		data, err := json.Marshal(a)
+
+		if err != nil {
+			return fmt.Errorf("Cannot encode data %v: %s", a.Question, err)
+		}
+
+		// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+
+		// Add an item to the BulkIndexer
+
+		err = bi.Add(
+			context.Background(),
+			esutil.BulkIndexerItem{
+				// Action field configures the operation to perform (index, create, delete, update)
+				Action: "index",
+
+				// DocumentID is the (optional) document ID
+				DocumentID: a.YoutubeURL + "-" + strconv.Itoa(order),
+
+				// Body is an `io.Reader` with the payload
+				Body: bytes.NewReader(data),
+
+				// OnSuccess is called for each successful operation
+				OnSuccess: func(ctx context.Context, item esutil.BulkIndexerItem, res esutil.BulkIndexerResponseItem) {
+					atomic.AddUint64(&countSuccessful, 1)
+				},
+
+				// OnFailure is called for each failed operation
+				OnFailure: func(ctx context.Context, item esutil.BulkIndexerItem, res esutil.BulkIndexerResponseItem, err error) {
+					if err != nil {
+						log.Printf("ERROR: %s", err)
+					} else {
+						log.Printf("ERROR: %s: %s", res.Error.Type, res.Error.Reason)
+					}
+				},
+			},
+		)
+		if err != nil {
+			return fmt.Errorf("Unexpected error: %s", err)
+		}
+		// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+	}
+
+	// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+	// Close the indexer
+	//
+	if err := bi.Close(context.Background()); err != nil {
+		return fmt.Errorf("Unexpected error: %s", err)
+	}
+	// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+	biStats := bi.Stats()
+
+	// Report the results: number of indexed docs, number of errors, duration, indexing rate
+	//
+	log.Println(strings.Repeat("▔", 65))
+
+	dur := time.Since(start)
+
+	if biStats.NumFailed > 0 {
+		return fmt.Errorf(
+			"Indexed [%s] documents with [%s] errors in %s (%s docs/sec)",
+			humanize.Comma(int64(biStats.NumFlushed)),
+			humanize.Comma(int64(biStats.NumFailed)),
+			dur.Truncate(time.Millisecond),
+			humanize.Comma(int64(1000.0/float64(dur/time.Millisecond)*float64(biStats.NumFlushed))),
+		)
+	} else {
+		log.Printf(
+			"Sucessfuly indexed [%s] documents in %s (%s docs/sec)",
+			humanize.Comma(int64(biStats.NumFlushed)),
+			dur.Truncate(time.Millisecond),
+			humanize.Comma(int64(1000.0/float64(dur/time.Millisecond)*float64(biStats.NumFlushed))),
+		)
+	}
+
+	return nil
+}
